+++ packages/vue/src/utils/index.ts
@@ -3,6 +3,7 @@ export { composeSlotClassName } from "./composeSlotClassName";
 export { mapPropsVariants } from "./mapPropsVariants";
 export { dataAttr } from "./dataAttr";
 export { createContext } from "./context";
+export { getPath, setPath } from "./path";
 export { motion, AnimatePresence } from "./motion";
 export { hasSlotComponent } from "./hasSlotComponent";
 export { warnDeprecatedVariant } from "./warnDeprecated";